104 u4 wip makefile.rej sndintrf.c.rej usrintrf.c.rej winmain.c.rej cpu.mak.rej
git-svn-id: https://mameplus.svn.sourceforge.net/svnroot/mameplus@371 e0215e0a-1e64-f442-9f1d-d2f61a7f4648

+++ trunk/mamep/src/vidhrdw/mario.c
@@ -7,10 +7,9 @@
 ***************************************************************************/
 
 #include "driver.h"
-#include "vidhrdw/generic.h"
 #include "state.h"
 
-static int gfx_bank, palette_bank;
+static UINT8 gfx_bank, palette_bank;
 
 static tilemap *bg_tilemap;
